+++ src/OutboxServiceProvider.php
@@ -5,6 +5,7 @@ namespace Webrek\Outbox;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Support\ServiceProvider;
 use Webrek\Outbox\Console\PruneCommand;
+use Webrek\Outbox\Console\RetryCommand;
 use Webrek\Outbox\Console\WorkCommand;
 use Webrek\Outbox\Contracts\Publisher;
 use Webrek\Outbox\Publishers\EventPublisher;
@@ -50,6 +51,7 @@ class OutboxServiceProvider extends ServiceProvider
             $this->commands([
                 WorkCommand::class,
                 PruneCommand::class,
+                RetryCommand::class,
             ]);
         }
     }